Factoring out utilities used in new plugin
